// client/src/autobindings/types.ts
export const AttackType = __t.enum("AttackType", {
  Sword: __t.unit(),
  Wand: __t.unit(),
  Knives: __t.unit(),
  Shield: __t.unit(),
  ThunderHorn: __t.unit(),
  AngelStaff: __t.unit(),
  Football: __t.unit(),
  Cards: __t.unit(),
  Dumbbell: __t.unit(),
  Garlic: __t.unit(),
  Volleyball: __t.unit(),
  Joint: __t.unit(),
});
export type AttackType = __Infer<typeof AttackType>;
// ...
export const ChosenUpgradeData = __t.object("ChosenUpgradeData", {
  chosenUpgradeId: __t.u32(),
  playerId: __t.u32(),
  get upgradeType() {
    return UpgradeType;
  },
  isAttackUpgrade: __t.bool(),
  value: __t.u32(),
  attackType: __t.u32(),
  damage: __t.u32(),
  cooldownRatio: __t.u32(),
  projectiles: __t.u32(),
  speed: __t.u32(),
  radius: __t.u32(),
  isNewAttack: __t.bool(),
});
export type ChosenUpgradeData = __Infer<typeof ChosenUpgradeData>;
// ...
export const UpgradeOptionData = __t.object("UpgradeOptionData", {
  upgradeId: __t.u32(),
  playerId: __t.u32(),
  upgradeIndex: __t.u32(),
  get upgradeType() {
    return UpgradeType;
  },
  isAttackUpgrade: __t.bool(),
  value: __t.u32(),
  attackType: __t.u32(),
  damage: __t.u32(),
  cooldownRatio: __t.u32(),
  projectiles: __t.u32(),
  speed: __t.u32(),
  radius: __t.u32(),
  isNewAttack: __t.bool(),
});
export type UpgradeOptionData = __Infer<typeof UpgradeOptionData>;

// The tagged union or sum type for the algebraic type `UpgradeType`.
export const UpgradeType = __t.enum("UpgradeType", {
  MaxHp: __t.unit(),
  HpRegen: __t.unit(),
  Speed: __t.unit(),
  Armor: __t.unit(),
  AttackSword: __t.unit(),
  AttackWand: __t.unit(),
  AttackKnives: __t.unit(),
  AttackShield: __t.unit(),
  AttackThunderHorn: __t.unit(),
  AttackAngelStaff: __t.unit(),
  AttackFootball: __t.unit(),
  AttackCards: __t.unit(),
  AttackDumbbell: __t.unit(),
  AttackGarlic: __t.unit(),
  AttackVolleyball: __t.unit(),
  AttackJoint: __t.unit(),
});
export type UpgradeType = __Infer<typeof UpgradeType>;